Create eventListenerWrapper.lua
Just simple sanity wrappers to stop Solar2D from warning about "key" events not being supported on iOS on Simulator, as well as to block calls to add or remove "key" event listeners on actual iOS devices.

Finally, also show a "loud" warning on Simulator if adding a "mouse" listener on iOS, because "mouse" events just don't work and there are no warning messages about it on the Simulator.